Use 64 bit parameter Id
Parameter id changed their size in the firwmware.
this is a patch to align the abi.

Note that the api is not aligned as a new bit needs to be set
for infrastructure parameters.

+++ components/cAVS/include/cAVS/DspFw/Common.hpp
@@ -60,7 +60,7 @@ namespace detail
 /** Unique trait to for parameterId. */
 struct ParameterIdTrait
 {
-    using RawType = uint32_t;
+    using RawType = uint64_t;
 };
 
 } // namespace detail
